// src/api/profile_api.ts
import axios from 'axios'

const instance = axios.create({
    baseURL: 'https://social-network.samuraijs.com/api/1.0/',
    withCredentials: true,
    headers: {
        'API-KEY': '54ca7025-138d-404e-bb10-b393f2399e64'
    }
})

export type ProfileResponseType = {
    userId: number
    lookingForAJob: boolean
    lookingForAJobDescription: string
    fullName: string
    contacts: {
        github: string
        vk: string
        facebook: string
        instagram: string
        twitter: string
        website: string
        youtube: string
        mainLink: string
    }
    photos: {
        small: string
        large: string
    }
}

export const profileAPI = {
    getProfile(userId: string) {
        return instance.get<ProfileResponseType>(`profile/${userId}`)
    }
}